Modification to implicit resolution of contravariant types.
Favor subtype-specificity over contrarivariant faux-specificity.
In a nutshell:

  trait Ord[-T] { }

  def f {
    implicit val Ord1: Ord[Iterable[Int]] = new Ord[Iterable[Int]] { override def toString = "Iterable" }
    implicit val Ord2: Ord[Seq[Int]]      = new Ord[Seq[Int]] { override def toString = "Seq" }
    implicit val Ord3: Ord[List[Int]]     = new Ord[List[Int]] { override def toString = "List" }

    println((implicitly[Ord[Iterable[Int]]], implicitly[Ord[Seq[Int]]], implicitly[Ord[List[Int]]]))
  }

In scala as it stands, this prints  (Iterable,Iterable,Iterable)
With this commit, it prints         (Iterable,Seq,List)

See files/run/specificity.scala for more examples.

With this change in place, it is easy to make Ordering and Equiv
contravariant as they deserve to be.